+++ src/app/modules/shared/services/json-converter.service.ts
@@ -350,6 +350,7 @@ export class JsonConverterService {
             character.level,
             character.version,
             character.originalVersion,
+            null,
             reaper,
             ultimatum,
             character.ancestralLegacies.nodes,